# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CivicPulse API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/me": {
      "get": {
        "summary": "Current User",
        "operationId": "current_user_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/complaints": {
      "post": {
        "summary": "Create Complaint",
        "operationId": "create_complaint_api_complaints_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Complaint"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Complaints",
        "operationId": "list_complaints_api_complaints_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Status"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintPage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/complaints/{id}": {
      "get": {
        "summary": "Get Complaint",
        "operationId": "get_complaint_api_complaints__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Complaint"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/complaints/{id}/status": {
      "patch": {
        "summary": "Update Status",
        "operationId": "update_status_api_complaints__id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Complaint"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "summary": "Get Stats",
        "operationId": "get_stats_api_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "headers": {
              "X-Cache": {
                "description": "Redis cache state",
                "schema": {
                  "type": "string",
                  "enum": [
                    "HIT",
                    "MISS"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/meta/providers": {
      "get": {
        "summary": "Get Providers",
        "operationId": "get_providers_api_meta_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersMeta"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "summary": "Choose Provider",
        "operationId": "choose_provider_api_meta_providers_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersMeta"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "string",
        "enum": [
          "water",
          "electricity",
          "sanitation",
          "roads",
          "streetlights",
          "other"
        ],
        "title": "Category"
      },
      "Complaint": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Contact"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "triaged_by": {
            "type": "string",
            "title": "Triaged By"
          },
          "triage_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triage Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "location",
          "reporter_contact",
          "category",
          "priority",
          "status",
          "ai_summary",
          "triaged_by",
          "triage_latency_ms",
          "created_at",
          "updated_at"
        ],
        "title": "Complaint"
      },
      "ComplaintCreate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Text"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Location"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Contact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "location"
        ],
        "title": "ComplaintCreate"
      },
      "ComplaintPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Complaint"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "ComplaintPage"
      },
      "CurrentUser": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "citizen",
              "operator"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "CurrentUser"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "Health"
      },
      "Priority": {
        "type": "string",
        "enum": [
          "high",
          "normal",
          "low"
        ],
        "title": "Priority"
      },
      "ProviderOption": {
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "groq",
              "ollama"
            ],
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "location": {
            "type": "string",
            "enum": [
              "hosted",
              "local"
            ],
            "title": "Location"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "model",
          "location",
          "available"
        ],
        "title": "ProviderOption"
      },
      "ProviderSelection": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "groq",
              "ollama"
            ],
            "title": "Provider"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "ProviderSelection"
      },
      "ProvidersMeta": {
        "properties": {
          "active_provider": {
            "type": "string",
            "title": "Active Provider"
          },
          "selected_provider": {
            "type": "string",
            "enum": [
              "groq",
              "ollama",
              "rules",
              "simulated"
            ],
            "title": "Selected Provider",
            "default": "rules"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/ProviderOption"
            },
            "type": "array",
            "title": "Options"
          },
          "recent": {
            "items": {
              "$ref": "#/components/schemas/TriageOutcome"
            },
            "type": "array",
            "title": "Recent"
          }
        },
        "type": "object",
        "required": [
          "active_provider",
          "recent"
        ],
        "title": "ProvidersMeta"
      },
      "Readiness": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "checks": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "status",
          "checks"
        ],
        "title": "Readiness"
      },
      "Stats": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Category"
          },
          "by_priority": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Priority"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Status"
          }
        },
        "type": "object",
        "required": [
          "total",
          "by_category",
          "by_priority",
          "by_status"
        ],
        "title": "Stats"
      },
      "Status": {
        "type": "string",
        "enum": [
          "open",
          "in_progress",
          "resolved",
          "rejected"
        ],
        "title": "Status"
      },
      "StatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusUpdate"
      },
      "TriageOutcome": {
        "properties": {
          "complaint_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complaint Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "fallback": {
            "type": "boolean",
            "title": "Fallback"
          },
          "error_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Class"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "complaint_id",
          "provider",
          "latency_ms",
          "fallback",
          "created_at"
        ],
        "title": "TriageOutcome"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "errors"
        ],
        "title": "ValidationErrorResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}